+++ test/dummy/config/routes.rb
@@ -2,6 +2,7 @@ Rails.application.routes.draw do
   get "up" => "rails/health#show", as: :rails_health_check
   root "pages#home"
   get "orders", to: "dashboards#show", as: :orders
+  get "gallery", to: "gallery#show", as: :gallery
   get "name", to: "pages#name", as: :the_name
   get "docs", to: "docs#index"
   get "docs/:slug", to: "docs#show", as: :doc, constraints: { slug: /[a-z0-9-]+/ }